refactor: ajustes de fundo do codigo
