+++ src/main/java/br/com/actionlabs/carboncalc/infra/config/AppConfig.java
@@ -16,7 +16,7 @@ public class AppConfig implements WebMvcConfigurer {
             auth ->
                 auth.requestMatchers(
                         "/login/**",
-                        "/open/**",
+                        "/calculator/**",
                         "/status/**",
                         "/swagger-ui/**",
                         "swagger-ui.html",
