+++ src/main/java/br/com/actionlabs/carboncalc/infra/repository/Test.java
@@ -1,0 +1,4 @@
+package br.com.actionlabs.carboncalc.infra.repository;
+
+public interface Test {
+}
